third_party/blink/web_tests/fast/scroll-behavior/overflow-scroll-animates.html

<!DOCTYPE html>
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<script src="../../resources/gesture-util.js"></script>
<script src="../../resources/testdriver.js"></script>
<script src="../../resources/testdriver-actions.js"></script>
<script src="../../resources/testdriver-vendor.js"></script>
<title>Tests that overflow scrolls are animated.</title>
<style>
  #container {
    overflow: scroll;
    width: 200px;
    height: 200px;
  }
  #content {
    width: 750px;
    height: 10000px;
    background-color: blue;
  }
</style>
<div id="container">
  <div id="content"></div>
</div>

<script>
if (!window.internals) {
  assert_unreached("This test must be run in the test harness");
}
internals.settings.setScrollAnimatorEnabled(true);

function smoothScrollPromise(container) {
  let firstScrollPosition = undefined;
  let lastScrollPosition = undefined;
  return new Promise((resolve, reject) => {
    const scrollendListener = () => {
      container.removeEventListener('scroll', scrollListener);
      if (lastScrollPosition == firstScrollPosition) {
        reject('expected multiple scroll events with a smooth scroll');
      } else {
        resolve();
      }
    }
    container.addEventListener('scrollend', scrollendListener, { once: true });
    const scrollListener = () => {
      if (firstScrollPosition == undefined) {
        firstScrollPosition = container.scrollTop;
      }
      lastScrollPosition = container.scrollTop;
    }
    container.addEventListener('scroll', scrollListener);
  });
}

promise_test(async () => {
  const container = document.getElementById("container");
  await waitForCompositorCommit();
  await waitForScrollReset(container);
  const animatedScroll = smoothScrollPromise(container);
  const pos = elementCenter(container);
  await mouseClick(pos.x, pos.y);
  await keyboardScroll("PageDown", container);
  await animatedScroll;
  assert_true(container.scrollTop > 0);
}, 'Overflow scrolls are animated');

</script>
